Improve extract feature
We were forcing the user to use a selected group of programs to extract
the downloaded files.

Now we added a functionality where users are asked which program they
want to use for that type of file. Once it registers, we save this
option and if we have this file extension again, we will use the
previous configuration and extract it automatically (if users used the
--extract option).