Implements download in chunks.
This commit implements the feature of downloading concurrently
(multi-threaded) in chunks. Downloaded chunks are then united
in a single file. This results in much better performance in
comparison of the traditional way, encouraged by the API, of
downloading chunks sequentially.

The implementation is backed by the `asyncio` standard library.

The script purportedly runs better on Python 3.9 or later, but a
best effort was made to backport it to Python 3.6 (please help
in test!).

Fixes #17.